refactor: Repository 레이어 제거
