--- src/main/java/wooteco/subway/repository/StationRepository.java
@@ -1,32 +1,0 @@
-package wooteco.subway.repository;
-
-import org.springframework.stereotype.Repository;
-import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.station.Station;
-
-import java.util.List;
-
-@Repository
-public class StationRepository {
-    private final StationDao stationDao;
-
-    public StationRepository(StationDao stationDao) {
-        this.stationDao = stationDao;
-    }
-
-    public void save(Station station) {
-        stationDao.insert(station);
-    }
-
-    public List<Station> findAll() {
-        return stationDao.findAll();
-    }
-
-    public void deleteById(Long id) {
-        stationDao.deleteById(id);
-    }
-
-    public boolean existByName(String name) {
-        return stationDao.existByName(name);
-    }
-}
