+++ src/main/java/wooteco/subway/service/StationService.java
@@ -1,43 +1,43 @@
 package wooteco.subway.service;
 
 import org.springframework.stereotype.Service;
+import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.station.Station;
 import wooteco.subway.dto.StationRequest;
 import wooteco.subway.dto.StationResponse;
-import wooteco.subway.repository.StationRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
 public class StationService {
-    private final StationRepository stationRepository;
+    private final StationDao stationDao;
 
-    public StationService(StationRepository stationRepository) {
-        this.stationRepository = stationRepository;
+    public StationService(StationDao stationDao) {
+        this.stationDao = stationDao;
     }
 
     public StationResponse save(StationRequest stationRequest) {
         Station station = stationRequest.toStation();
         checkDuplication(station);
-        stationRepository.save(station);
+        stationDao.insert(station);
         return new StationResponse(station.getId(), station.getName());
     }
 
     private void checkDuplication(Station station) {
-        if (stationRepository.existByName(station.getName())) {
+        if (stationDao.existByName(station.getName())) {
             throw new IllegalArgumentException("이미 존재하는 역 이름입니다.");
         }
     }
 
     public List<StationResponse> findAll() {
-        List<Station> stations = stationRepository.findAll();
+        List<Station> stations = stationDao.findAll();
         return stations.stream()
                 .map(station -> new StationResponse(station.getId(), station.getName()))
                 .collect(Collectors.toList());
     }
 
     public void delete(Long id) {
-        stationRepository.deleteById(id);
+        stationDao.deleteById(id);
     }
 }
